refactor: hoist Codec into utz-common and type decompress with it
The encoder (utz-encode) and reader dispatched on the same wire byte
with different vocabularies: an enum on one side, a raw u8 on the
other. Codec now lives in utz-common (no_std, zero deps; its alloc
samplers move behind a default-on alloc feature so the enum reaches
the bare core rung), utz-encode re-exports it, and utz re-exports it
and takes it in decompress's signature. The dispatch match is now
exhaustive over the enum - compiled-in codecs decode, complementary
cfg arms return CodecNotCompiledIn - and the wire byte is parsed once
at the finder boundary via Codec::from_byte, with unknown bytes
rejected there. Error variants keep raw-byte payloads so unknown
bytes stay representable. window_sweep drops its codec-as-u8 cast.

+++ utz-build/src/cmd/window_sweep.rs
@@ -120,8 +120,7 @@ pub fn run(a: &Args) -> utz_build::Result<()> {
         // decode through the shipped path; peak-RAM model: decoded + window + state
         let row =
             |name: &str, window: usize, codec: Codec, blob: Vec<u8>| -> utz_build::Result<()> {
-                let (out, peak, ms) =
-                    measure(|| utz::decompress::decompress(codec as u8, raw, &blob));
+                let (out, peak, ms) = measure(|| utz::decompress::decompress(codec, raw, &blob));
                 let out = out.map_err(|e| Error::Msg(format!("{name} decode: {e:?}")))?;
                 ensure!(
                     out == payload,
